basic movment working
